* added a status page to the MRC * created a new, simpler scrubber tool * fixed the checkObject method in the OSD * fixes for the RandomAccessFile * fixed the renew operation in the MRC
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@202 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ servers/test/org/xtreemfs/test/io/ByteMapperTest.java
@@ -95,7 +95,7 @@ public class ByteMapperTest extends TestCase{
     }
     
     class TestObjectStore implements ObjectStore{
-        public ReusableBuffer readObject(long objectNo, long offset, long length){
+        public ReusableBuffer readObject(long objectNo, int offset, int length){
             String content = "Hallo World";
             return ReusableBuffer.wrap(content.substring((int) offset, (int) (offset+length)).getBytes());
         }
@@ -107,7 +107,7 @@ public class ByteMapperTest extends TestCase{
     }
     
     class EmptyObjectStore implements ObjectStore{
-        public ReusableBuffer readObject(long objectNo, long offset, long length){
+        public ReusableBuffer readObject(long objectNo, int offset, int length){
             return ReusableBuffer.wrap("".getBytes());
         }
